requirements.txt
anyio==3.3.4
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cycler==0.10.0
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.0
h11==0.12.0
httptools==0.2.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
nltk==3.6.5
numpy==1.21.3
orjson==3.6.4
pandas==1.3.4
Pillow==8.4.0
pydantic==1.8.2
pyparsing==3.0.3
python-dateutil==2.8.2
python-dotenv==0.19.1
python-multipart==0.0.5
pytz==2021.3
PyYAML==5.4.1
regex==2021.10.23
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
sklearn==0.0
sniffio==1.2.0
starlette==0.16.0
threadpoolctl==3.0.0
tqdm==4.62.3
typing-extensions==3.10.0.2
ujson==4.2.0
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.0
